scoreboard removed, powershell and cmd support

+++ build.py
@@ -1,7 +1,9 @@
 import sys
 import os
-import json
 import subprocess
+
+def is_powershell():
+    return 'PSModulePath' in os.environ
 
 def installdeps():
     print("Setting up virtual environment...")
@@ -16,22 +18,20 @@ def installdeps():
 def build():
     print("Building executable...")
     if sys.platform=="win32":
-        subprocess.run([sys.executable,"-m","PyInstaller","-F","src/termadventure.py","--add-data","termadventure/data/scoreboard.json;data"],text=True,check=True)
+        subprocess.run([sys.executable,"-m","PyInstaller","-F","src/termadventure.py"],text=True,check=True)
     elif sys.platform=="linux":
-        subprocess.run(["./venv/bin/python","-m","PyInstaller","-F","src/termadventure.py","--add-data","termadventure/data/scoreboard.json;data"],text=True,check=True)
+        subprocess.run(["./venv/bin/python","-m","PyInstaller","-F","src/termadventure.py"],text=True,check=True)
     print("Cleaning up...")
     if sys.platform=="win32":
-        subprocess.run(["powershell.exe", "-Command", "Remove-Item","-Recurse","build"],text=True,check=True)
-        subprocess.run(["powershell.exe", "-Command", "Remove-Item","termadventure.spec"],text=True,check=True)
+        if is_powershell():
+            subprocess.run(["powershell.exe", "-Command", "Remove-Item","-Recurse","build"],text=True,check=True)
+            subprocess.run(["powershell.exe", "-Command", "Remove-Item","termadventure.spec"],text=True,check=True)
+        else:
+            subprocess.run(["rd", "/s", "/q", "build"], check=True)
+            subprocess.run(["del","termadventure.spec"], check=True)
     elif sys.platform=="linux":
         subprocess.run(["rm","-rf","build"],text=True,check=True)
         subprocess.run(["rm","termadventure.spec"],text=True,check=True)
-    print("Creating json scoreboard...")
-    os.mkdir("data")
-    filepath = os.path.join("data", "scoreboard.json")
-    jsondata=[]
-    with open(filepath, 'w') as f:
-        json.dump(jsondata, f, indent=2)
         
 if __name__ == "__main__":
     installdeps()
